Consolidate enums to single model
