+++ rocks.kfs.Intacct/Utils/TransactionHelpers.cs
@@ -22,6 +22,7 @@ using Rock;
 using Rock.Data;
 using Rock.Model;
 using Rock.Web.Cache;
+using rocks.kfs.Intacct.Enums;
 using KFSConst = rocks.kfs.Intacct.SystemGuid;
 
 namespace rocks.kfs.Intacct.Utils
